requirements.txt
antlr4-python3-runtime==4.13.0
graphviz==0.20.1
tk==0.1.0
